Updating TORQUE's license to reflect the fact that it isn't OpenPBS, where it is now maintained, etc.

+++ src/lib/Libattr/attr_fn_arst.c
@@ -569,7 +569,7 @@ int set_arst(
       return(PBSE_INTERNAL);
       }
 
-    need = (int)sizeof(struct array_strings) + (j - 1) * sizeof(char *);
+    need = sizeof(struct array_strings) + (j - 1) * sizeof(char *);
 
     pas = (struct array_strings *)malloc((size_t)need);
